Add dynamic completion for release names
This commit adds dynamic completion for the commands
   helm status <TAB>
   helm uninstall <TAB>
   helm history <TAB>
   helm test run <TAB>
   helm upgrade <TAB>
   helm get <subcommand> <TAB>
   helm rollback <TAB>

Support for override flags in completion is included for
such dynamic completion.

Finally, flags sometimes can be used with an = sign, such as
--kube-context=prod.
In this case, the variable ${flagname} retains the = sign as part of the
flag name. However, in zsh completion, an = sign cannot be part of an
index of the associative array 'flaghash' or else it causes an error.

This commit also strips the = sign out when using ${flagname} as an index
for zsh.

Note that this is not a big deal since flaghash is not actually used
anywhere in Helm completion. I believe it is made available by the
Cobra framework in case some completions choose to use it.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>